requirements.txt
blinker==1.9.0
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.1.8
cryptography==46.0.1
Flask==3.1.2
gTTS==2.5.4
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
lxml==6.0.2
MarkupSafe==3.0.2
packaging==25.0
pdfminer.six==20250506
pdfplumber==0.11.7
pillow==11.3.0
pycparser==2.23
pydub==0.25.1
PyMuPDF==1.26.4
pypdfium2==4.30.0
pytesseract==0.3.13
python-docx==1.2.0
pyttsx3==2.99
requests==2.32.5
SpeechRecognition==3.14.3
typing_extensions==4.15.0
urllib3==2.5.0
Werkzeug==3.1.3
# Additional libraries for enhanced functionality
ffmpeg-python==0.2.0
opencv-python==4.10.0.84
reportlab==4.2.2
img2pdf==0.5.1
